Fixes to #304 and #311
