+++ internal/app/wwctl/configure/dhcp/main.go
@@ -6,6 +6,5 @@ import (
 )
 
 func CobraRunE(cmd *cobra.Command, args []string) error {
-	return configure.Configure("DHCP", setShow)
+	return configure.Dhcp()
 }
-
