+++ internal/app/wwctl/configure/dhcp/root.go
@@ -13,11 +13,9 @@ var (
 			"in the warewulf.conf file.",
 		RunE: CobraRunE,
 	}
-	setShow bool
 )
 
 func init() {
-	baseCmd.PersistentFlags().BoolVarP(&setShow, "show", "s", false, "Show configuration (don't update)")
 }
 
 // GetRootCommand returns the root cobra.Command for the application.
