+++ internal/app/wwctl/configure/nfs/main.go
@@ -6,5 +6,5 @@ import (
 )
 
 func CobraRunE(cmd *cobra.Command, args []string) error {
-	return configure.Configure("NFS", setShow)
+	return configure.NFS()
 }
