+++ internal/app/wwctl/configure/nfs/root.go
@@ -11,11 +11,9 @@ var (
 			"configure NFS as per the configuration in the warewulf.conf file.",
 		RunE: CobraRunE,
 	}
-	setShow bool
 )
 
 func init() {
-	baseCmd.PersistentFlags().BoolVarP(&setShow, "show", "s", false, "Show configuration (don't update)")
 }
 
 // GetRootCommand returns the root cobra.Command for the application.
