+++ internal/app/wwctl/configure/root.go
@@ -1,13 +1,10 @@
 package configure
 
 import (
-	"os"
-
 	"github.com/hpcng/warewulf/internal/app/wwctl/configure/dhcp"
 	"github.com/hpcng/warewulf/internal/app/wwctl/configure/nfs"
 	"github.com/hpcng/warewulf/internal/app/wwctl/configure/ssh"
 	"github.com/hpcng/warewulf/internal/app/wwctl/configure/tftp"
-	"github.com/hpcng/warewulf/internal/pkg/configure"
 	"github.com/spf13/cobra"
 )
 
@@ -34,21 +31,4 @@ func init() {
 // GetRootCommand returns the root cobra.Command for the application.
 func GetCommand() *cobra.Command {
 	return baseCmd
-}
-
-func CobraRunE(cmd *cobra.Command, args []string) error {
-	var err error
-	if allFunctions {
-		for _, s := range [5]string{"DHPC", "NFS", "SSH", "TFTP"} {
-			err = configure.Configure(s, false)
-			if err != nil {
-				os.Exit(1)
-			}
-		}
-	} else {
-		_ = cmd.Help()
-		os.Exit(0)
-	}
-
-	return nil
 }
