+++ internal/pkg/configure/dhcp.go
@@ -4,18 +4,18 @@ import (
 	"fmt"
 	"os"
 
-  "github.com/hpcng/warewulf/internal/pkg/overlay"
+	"github.com/hpcng/warewulf/internal/pkg/overlay"
 	"github.com/hpcng/warewulf/internal/pkg/util"
 	"github.com/hpcng/warewulf/internal/pkg/warewulfconf"
 	"github.com/hpcng/warewulf/internal/pkg/wwlog"
 	"github.com/pkg/errors"
 )
+
 /*
-Configures the dhcpd server, when show is set to false, else the 
+Configures the dhcpd server, when show is set to false, else the
 dhcp configuration is checked.
 */
-func configureDHCP(show bool) error {
-
+func Dhcp() error {
 
 	controller, err := warewulfconf.New()
 	if err != nil {
@@ -37,18 +37,16 @@ func configureDHCP(show bool) error {
 		wwlog.Printf(wwlog.ERROR, "Configuration is not defined: `dhcpd range end`\n")
 		os.Exit(1)
 	}
-	if !show {
 
-		err = overlay.BuildHostOverlay()
-    if err != nil {
-			wwlog.Printf(wwlog.WARN, "host overlay could not be built: %s\n", err)
-		}
-		fmt.Printf("Enabling and restarting the DHCP services\n")
-		err := util.SystemdStart(controller.Dhcp.SystemdName)
-		if err != nil {
-			return errors.Wrap(err, "failed to start")
-		}
+	err = overlay.BuildHostOverlay()
+	if err != nil {
+		wwlog.Printf(wwlog.WARN, "host overlay could not be built: %s\n", err)
 	}
-	return nil
+	fmt.Printf("Enabling and restarting the DHCP services\n")
+	err = util.SystemdStart(controller.Dhcp.SystemdName)
+	if err != nil {
+		return errors.Wrap(err, "failed to start")
+	}
 
+	return nil
 }
